feature: Journal Prototype
